minimal_requirements.txt
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2
pytest==7.4.0